requirements.txt
click==8.1.7
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.0